Converted client digging packet
